# cedar-resource-server-application/src/main/resources/assets/swagger-api/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Better Metadata - Better Science",
    "version" : "2.8.3",
    "title" : "CEDAR Resource Server API"
  },
...
  "schemes" : [ "https" ],
  "paths" : {
    "/categories" : {
      "get" : {
        "tags" : [ "Categories" ],
        "summary" : "Get all categories",
        "description" : "Get the list of all categories.",
        "operationId" : "getAllCategories",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "limit",
          "in" : "query",
          "description" : "Paging limit",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "offset",
          "in" : "query",
          "description" : "Paging offset",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "A category",
            "schema" : {
              "$ref" : "#/definitions/Category"
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "Categories" ],
        "summary" : "Create a category",
        "description" : "Create a category.",
        "operationId" : "createCategory",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "category",
          "description" : "The category to be created",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Category"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Category"
            }
          },
          "201" : {
            "description" : "A category",
            "schema" : {
              "$ref" : "#/definitions/Category"
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
    "/categories/root" : {
      "get" : {
        "tags" : [ "Categories" ],
        "summary" : "Get root category",
        "description" : "Get root category.",
        "operationId" : "findRootCategory",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "A category",
            "schema" : {
              "$ref" : "#/definitions/Category"
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
    "/categories/tree" : {
      "get" : {
        "tags" : [ "Categories" ],
        "summary" : "Get category tree",
        "description" : "Get category tree.",
        "operationId" : "findCategoryTree",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "category_id",
          "in" : "path",
          "description" : "Category identifier. Example: https://repo.metadatacenter.org/categories/8bc64ab5-df6b-48c8-8c61-6c016245918e",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "A category",
            "schema" : {
              "$ref" : "#/definitions/Category"
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
    "/categories/{category_id}" : {
      "get" : {
        "tags" : [ "Categories" ],
        "summary" : "Get a category",
        "description" : "Get a category.",
        "operationId" : "findCategory",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "category_id",
          "in" : "path",
          "description" : "Category identifier. Example: https://repo.metadatacenter.org/categories/8bc64ab5-df6b-48c8-8c61-6c016245918e",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "A category",
            "schema" : {
              "$ref" : "#/definitions/Category"
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Categories" ],
        "summary" : "Update a category",
        "description" : "Update a category.",
        "operationId" : "updateCategory",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "category_id",
          "in" : "path",
          "description" : "Category identifier. Example: https://repo.metadatacenter.org/categories/8bc64ab5-df6b-48c8-8c61-6c016245918e",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "category",
          "description" : "The category to be updated",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Category"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "A category",
            "schema" : {
              "$ref" : "#/definitions/Category"
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Categories" ],
        "summary" : "Delete a category",
        "description" : "Delete a category.",
        "operationId" : "deleteCategory",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "category_id",
          "in" : "path",
          "description" : "Category identifier. Example: https://repo.metadatacenter.org/categories/8bc64ab5-df6b-48c8-8c61-6c016245918e",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Successful operation (no content)"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
...
  },
  "securityDefinitions" : {
    "api_key" : {
      "description" : "Your CEDAR API key, using the format 'apiKey <yourApiKey>'",
      "type" : "apiKey",
      "name" : "Authorization",
      "in" : "header"
    }
  },
  "definitions" : {
    "Category" : {
      "type" : "object",
      "properties" : {
        "@id" : {
          "type" : "string",
          "description" : "Unique URL identifier representing a specific category."
        },
        "parentCategoryId" : {
          "type" : "string",
          "description" : "Unique URL identifier representing a specific parent category."
        },
        "schema:description" : {
          "type" : "string",
          "description" : "Description of the category."
        },
        "schema:identifier" : {
          "type" : "string",
          "description" : "Identifier, used for identifying this object in outside-to-CEDAR systems."
        },
        "schema:name" : {
          "type" : "string",
          "description" : "Name of the category."
        }
      },
      "description" : "A CEDAR category."
    },
...
  }
}